backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
langchain==0.1.0
langchain-community==0.0.10
ollama==0.1.6
pydantic==2.5.0
gunicorn==21.2.0
lxml==4.9.3
python-dateutil==2.8.2

